Extract RDW parsing primitives into a dedicated microcrate
