+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "copybook-overflow",
     "copybook-fixed",
     "copybook-rdw",
+    "copybook-rdw-primitives",
     "copybook-record-io",
     "copybook-overpunch",
     "copybook-determinism",
@@ -57,6 +58,7 @@ copybook-utils = { version = "=0.4.3", path = "copybook-utils" }
 copybook-overflow = { version = "=0.4.3", path = "copybook-overflow" }
 copybook-fixed = { version = "=0.4.3", path = "copybook-fixed" }
 copybook-rdw = { version = "=0.4.3", path = "copybook-rdw" }
+copybook-rdw-primitives = { version = "=0.4.3", path = "copybook-rdw-primitives" }
 copybook-record-io = { version = "=0.4.3", path = "copybook-record-io" }
 copybook-overpunch = { version = "=0.4.3", path = "copybook-overpunch" }
 copybook-determinism = { version = "=0.4.3", path = "copybook-determinism" }
