+++ copybook-rdw-primitives/README.md
@@ -1,0 +1,3 @@
+# copybook-rdw-primitives
+
+Low-level RDW header and buffered parsing primitives extracted from `copybook-rdw`.
